pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "faststats"
version = "0.0.0"
authors = [
    {name = "Andrew Landau", email = "andrew+tyler+landau+getridofthisanddtheplusses@gmail.com"},
]
description = "A simple package to support fast statistics."
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["numpy", "numba"]
license = {file = "LICENSE"}
keywords = ["statistics", "numpy", "fast", "performance", "numba", "data", "analysis"]   
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[tool.hatch.build.hooks.custom]
path = "tools/hooks.py"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
benchmark = [
    "tabulate",
]
dev = [
    "black",
    "tools/generators",
    "pyyaml",
]
# TODO: Fix this. I want the install hook to be included, but not the generator code. 
# I only want generator code to be included when the user requests a [dev] install.
# Don't think the above is correct....

[project.urls]
Homepage = "https://github.com/landoskape/faststats"

# TODO: Fix this. I want the install hook to be included, but not the generator code. 
# I only want generator code to be included when the user requests a [dev] install.
[tool.hatch.build]
include = ["faststats", "tools/build/hooks.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=faststats --cov-report=xml --cov-report=term-missing"